backend/requirements.txt
# API
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9
pydantic-settings==2.1.0

# Database
psycopg2-binary==2.9.9

# Utils
python-dotenv==1.0.1
httpx==0.26.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3

# RAG and LLM (Phase 4)
haystack-ai==2.20.0
pgvector-haystack==5.4.0
sentence-transformers==3.3.0
torch==2.5.1
pandas==2.2.3  # Required by pgvector-haystack

# Future LLM integrations
# ollama-haystack
# openai
# anthropic
